+++ src/main/java/ru/develonica/service/SizeHandler.java
@@ -14,5 +14,5 @@ public interface SizeHandler {
      * @return Size value.
      * @throws Exception Exception.
      */
-    double activate(File destination) throws Exception;
+    long activate(File destination) throws Exception;
 }
